+++ packages/conc/changelog.md
@@ -6,6 +6,7 @@
 * Add `Semaphore`, abstracting `QSem` or `TSem`.
 * Add `timeoutStop`, a variant of `timeout` that calls `Polysemy.Resume.stop`.
 * Add `Lock`, a mutex effect.
+* Add `withAsyncTrigger`, a variant of `withAsync` that requires the async action to signal the sync action to start.
 
 # 0.9.0.0